app/requirements.txt
langchain-community==0.2.4
langchain-google-genai==1.0.6
langchain-text-splitters==0.2.1
chromadb==0.4.17
langchainhub==0.1.18
streamlit==1.35.0
beautifulsoup4==4.12.3
html2text==2024.2.26